feat: wip here we add all the draft7 schema to maek sure we dont
crash when we perform rendering. In this imm. commit we play with
having a renderer that just does json schema, but we might need
to pivot to rendering content descriptors and json schema as separate
things there are scenarios where the parameters and perhaps
results squash into the parameters and it makes sense to display
the content descriptor as if it were on the schema data itself

So taking that into account it means we can't really have a json
schema rendering aspect outside of the content descriptor because
the data is really one object coupled vs separate entities.

In an object we want to display the full object with it's additional
schema details, when in the parameter if it is simple we want
to just display the parameter name for the entry and description
as any additional information would be confusing potentially to
display.