feat: wip renderer now works for most cases except the anyof,oneof,cases
it's missing pattern and requirement additional data and custom renderer
things. For the most part those aren't necessary from what I can tell
for the features we're supporting.

Next steps are to wire this so it live renders the docs after completing
the minimal oneof/anyof so the schema doesn't break while rendering

needs to write to disk and needs to write final output to ast

currently takes just dereffed method and will produce the proper
output as in the widget library mdx components which is a live render